Initial unit tests for bot config cred provider
This adds an initial set of unit tests for the
`BotConfigurationEndpointServiceCredentialProvider` class.

NOTE: because of the direct ties to the file system for some methods,
this is somewhat difficult to test. In fact one test for the `Load`
method is failing right now because there's no way to control when the
.bot file is there for one test and not for another.